Add more tests; Fix code standards

+++ test/exceptions_test/bad_rest_file_malformed/run.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-
-_SCRIPT_DIR=$(dirname ${0})
+cd $(dirname ${0})
+_SCRIPT_DIR=$(pwd)
 _TMP="${_SCRIPT_DIR}/.tmp$(date +%s)"
 
 . ${_SCRIPT_DIR}/../../../ci/common.sh
@@ -10,7 +10,6 @@ cd ${_SCRIPT_DIR}
 restafari -s localhost -p ${TEST_SERVER_PORT} file.rest >> ${_TMP}
 cat ${_TMP}
 
-[[ -n "$(grep Trackback $_TMP)" ]] && exit 1
-[[ -n "$(grep "malformed" $_TMP)" ]] && exit 0
-
-rm -f ${_TMP}
+[[ -n "$(grep Trackback $_TMP)" ]]   && rm -f ${_TMP} && exit 1
+[[ -n "$(grep "malformed" $_TMP)" ]] && rm -f ${_TMP} && exit 0
+exit 1
